% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{basesurv.glmnet}
\alias{basesurv.glmnet}
\title{Breslow estimator for baseline hazard}
\usage{
basesurv.glmnet(time, event, lp, times.eval = NULL, centered = FALSE)
}
\value{
list containing cumulative base hazard
}
\description{
Derived from \code{peperr:::basesurv} and \code{gbm::basehaz.gbm}.
}
\keyword{internal}

