% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_5_calibrate_external_utils.R
\name{calibrate_external_surv_prob_true}
\alias{calibrate_external_surv_prob_true}
\title{Compute Kaplan-Meier estimated survival probabilities for external calibration}
\usage{
calibrate_external_surv_prob_true(pred_prob, grp, time_new, event_new,
  pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier estimated survival probabilities for external calibration
}
\keyword{internal}
