% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_utils.R
\name{calibrate_surv_prob_true}
\alias{calibrate_surv_prob_true}
\title{Compute Kaplan-Meier estimated survival probabilities for calibration}
\usage{
calibrate_surv_prob_true(pred_prob, grp, time, event, pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier estimated survival probabilities for calibration
}
\keyword{internal}
