% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hddnom-datalist.R
\docType{data}
\name{gbmsurv}
\alias{gbmsurv}
\title{Some TCGA dataset (mRNA expression and survival time)}
\usage{
data(gbmsurv)
}
\description{
Some TCGA dataset (mRNA expression and survival time)
}
\details{
Complimentary with the SMART clinical data, this is
the molecular-level data.
}
\examples{
# data(gbmsurv)
# str(gbmsurv)
}

