% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.nomogram.R
\name{glmnet.basesurv}
\alias{glmnet.basesurv}
\title{Breslow baseline hazard estimator}
\usage{
glmnet.basesurv(time, event, lp, times.eval = NULL, centered = FALSE)
}
\value{
list containing cumulative base hazard
}
\description{
Derived from \code{peperr:::basesurv} and \code{gbm::basehaz.gbm}.
}
\keyword{internal}

