% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{glmnet.calibrate.internal.pred}
\alias{glmnet.calibrate.internal.pred}
\title{Compute predicted survival probabilities for calibration}
\usage{
glmnet.calibrate.internal.pred(x_tr, x_te, y_tr, y_te, alpha, lambda, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute predicted survival probabilities for calibration
}
\keyword{internal}

