% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{glmnet.calibrate.internal.true}
\alias{glmnet.calibrate.internal.true}
\title{Compute Kaplan-Meier estimated survival probabilities for calibration}
\usage{
glmnet.calibrate.internal.true(pred_prob, grp, time, event, pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier estimated survival probabilities for calibration
}
\keyword{internal}

