% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.hdnom.external.calibrate.R
\name{glmnet.external.calibrate.internal.pred}
\alias{glmnet.external.calibrate.internal.pred}
\title{Compute glmnet Predicted Survival Probabilities for External Calibration}
\usage{
glmnet.external.calibrate.internal.pred(object, x_tr, x_te, y_tr, alpha, lambda,
  pen.factor, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute glmnet Predicted Survival Probabilities for External Calibration
}
\keyword{internal}

