% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.hdnom.external.validate.R
\name{glmnet.external.validate.internal}
\alias{glmnet.external.validate.internal}
\title{Compute External Validation Measures for glmnet Objects}
\usage{
glmnet.external.validate.internal(x_tr, x_te, y_tr, y_te, alpha, lambda,
  pen.factor, tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute External Validation Measures for glmnet Objects
}
\keyword{internal}

