% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.models.R
\name{glmnet.tune.alpha}
\alias{glmnet.tune.alpha}
\title{Automatic alpha tuning function by k-fold cross-validation}
\usage{
glmnet.tune.alpha(..., alphas, seed)
}
\value{
best model object and best alpha
}
\description{
Automatic alpha tuning function by k-fold cross-validation
}
\keyword{internal}

