% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.validate.R
\name{glmnet.validate.internal}
\alias{glmnet.validate.internal}
\title{Compute validation measures for glmnet objects}
\usage{
glmnet.validate.internal(samp_idx, nomogram, x, time, event, s)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute validation measures for glmnet objects
}
\keyword{internal}

