% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.validate.R
\name{glmnet.validate}
\alias{glmnet.validate}
\title{Validate Cox models fitted with glmnet using time-dependent AUC}
\usage{
glmnet.validate(nomogram, x, time, event, s = nomogram$s,
  method = c("bootstrap", "cv", "repeated.cv"), B = NULL, nfolds = NULL,
  rep.times = NULL)
}
\arguments{
\item{nomogram}{object returned by \code{\link{glmnet.nomogram}}.}

\item{x}{Matrix of training data used for the \code{glmnet} object;
on which to run the validation.}

\item{time}{Survival time.
Must be of the same length with the number of rows as \code{x}.}

\item{event}{Status indicator, normally 0 = alive, 1 = dead.
Must be of the same length with the number of rows as \code{x}.}

\item{s}{Value of the penalty parameter lambda in glmnet. By default,
we use the same penalty parameter as the nomogram object.}

\item{method}{Validation method. "bootstrap", "cv", or "repeated.cv".}

\item{B}{Number of repetitions for bootstrap.}

\item{nfolds}{Number of folds for cross-validation and
repeated cross-validation.}

\item{rep.times}{Number of repeated times for repeated cross-validation.}
}
\description{
Validate Cox models fitted with glmnet using time-dependent AUC
}
\examples{
set.seed(1002)
}

