% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_5_calibrate_external_utils.R
\name{glmnet_calibrate_external_surv_prob_pred}
\alias{glmnet_calibrate_external_surv_prob_pred}
\title{Compute glmnet predicted survival probabilities for external calibration}
\usage{
glmnet_calibrate_external_surv_prob_pred(object, x_tr, x_te, y_tr, alpha,
  lambda, pen.factor, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute glmnet predicted survival probabilities for external calibration
}
\keyword{internal}
