% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_utils.R
\name{glmnet_calibrate_surv_prob_pred}
\alias{glmnet_calibrate_surv_prob_pred}
\title{Compute glmnet predicted survival probabilities for calibration}
\usage{
glmnet_calibrate_surv_prob_pred(
  x_tr,
  x_te,
  y_tr,
  alpha,
  lambda,
  pen.factor,
  pred.at,
  cox.ties
)
}
\value{
list containing predicted survival probability
}
\description{
Compute glmnet predicted survival probabilities for calibration
}
\keyword{internal}
