% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_survial_curves.R
\name{glmnet_survcurve}
\alias{glmnet_survcurve}
\title{Survival curve prediction for glmnet objects}
\usage{
glmnet_survcurve(object, time, event, x, survtime)
}
\arguments{
\item{object}{\code{glmnet} model object}

\item{time}{Survival time}

\item{event}{Status indicator}

\item{x}{Predictor matrix}

\item{survtime}{Survival time to evaluate}
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\examples{
NULL
}
