% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_model_tune.R
\name{glmnet_tune_alpha}
\alias{glmnet_tune_alpha}
\title{Automatic alpha tuning function by k-fold cross-validation}
\usage{
glmnet_tune_alpha(..., alphas, seed, parallel)
}
\value{
best model object and best alpha
}
\description{
Automatic alpha tuning function by k-fold cross-validation
}
\keyword{internal}
