% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_5_validate_external_utils.R
\name{glmnet_validate_external_tauc}
\alias{glmnet_validate_external_tauc}
\title{Compute external validation measures for glmnet objects}
\usage{
glmnet_validate_external_tauc(
  object,
  x_tr,
  x_te,
  y_tr,
  y_te,
  tauc.type,
  tauc.time
)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute external validation measures for glmnet objects
}
\keyword{internal}
