% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.models.R
\name{hdcox.enet}
\alias{hdcox.enet}
\title{Elastic-net model selection for high-dimensional Cox models}
\usage{
hdcox.enet(x, y, nfolds = 5L, alphas = seq(0.05, 0.95, 0.05),
  rule = c("lambda.min", "lambda.1se"), seed = 1001)
}
\arguments{
\item{x}{data matrix}

\item{y}{response matrix made by \code{\link[survival]{Surv}}}

\item{nfolds}{fold numbers of cross-validation}

\item{alphas}{alphas to tune in \code{\link[glmnet]{cv.glmnet}}}

\item{rule}{Model selection criterion, \code{"lambda.min"} or
\code{"lambda.1se"}. See \code{\link[glmnet]{cv.glmnet}}
for details.}

\item{seed}{a random seed for cross-validation fold division}
}
\description{
Automatic elastic-net model selection for high-dimensional
Cox models, evaluated by penalized partial-likelihood.
}
\examples{
library("glmnet")
library("survival")
library("rms")
library("doParallel")
registerDoParallel(detectCores())

# Load imputed SMART data
data("smart")
x = as.matrix(smart[, -c(1, 2)])
time = smart$TEVENT
event = smart$EVENT
y = Surv(time, event)

# Fit Cox model by adaptive elastic-net penalization
enetfit = hdcox.enet(x, y, nfolds = 10, rule = 'lambda.1se', seed = 11)

# Prepare data for glmnet.nomogram
x.df = as.data.frame(x)
dd = datadist(x.df)
options(datadist = "dd")

# Generate glmnet.nomogram objects and draw nomogram
nom = glmnet.nomogram(enetfit$enet_all, x, time, event, x.df,
                      s = enetfit$best_lambda_enet, pred.at = 365 * 2,
                      funlabel = "2-Year Overall Survival Probability")

plot(nom)
}

