% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.models.R
\name{hdcox.lasso}
\alias{hdcox.lasso}
\title{Lasso model selection for high-dimensional Cox models}
\usage{
hdcox.lasso(x, y, nfolds = 5L, alphas = seq(0.05, 0.95, 0.05))
}
\arguments{
\item{x}{data matrix}

\item{y}{response matrix made by \code{\link[survival]{Surv}}}
}
\description{
Automatic lasso model selection for high-dimensional
Cox models, evaluated by time-dependent AUC (tAUC).
}
\examples{
set.seed(1001)
}

