% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.models-mcp.R
\name{hdcox.mcp}
\alias{hdcox.mcp}
\title{MCP model selection for high-dimensional Cox models}
\usage{
hdcox.mcp(x, y, nfolds = 5L, gammas = c(1.01, 1.7, 3, 100), seed = 1001,
  trace = FALSE, parallel = FALSE)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response matrix made by \code{\link[survival]{Surv}}.}

\item{nfolds}{Fold numbers of cross-validation.}

\item{gammas}{Gammas to tune in \code{\link[ncvreg]{cv.ncvsurv}}.}

\item{seed}{A random seed for cross-validation fold division.}

\item{trace}{Output the cross-validation parameter tuning
progress or not. Default is \code{FALSE}.}
}
\description{
Automatic MCP model selection for high-dimensional
Cox models, evaluated by penalized partial-likelihood.
}
\examples{
library("ncvreg")
library("survival")
library("rms")

# Load imputed SMART data
data("smart")
x = as.matrix(smart[, -c(1, 2)])
time = smart$TEVENT
event = smart$EVENT
y = Surv(time, event)

# Fit Cox model by MCP penalization
mcpfit = hdcox.mcp(x, y, nfolds = 3, gammas = c(1.7, 3), trace = TRUE)

# Prepare data for hdnom.nomogram
x.df = as.data.frame(x)
dd = datadist(x.df)
options(datadist = "dd")

# Generate hdnom.nomogram objects and plot nomogram
nom = hdnom.nomogram(mcpfit$mcp_model, x, time, event, x.df,
                     gamma = mcpfit$mcp_best_gamma,
                     lambda = mcpfit$mcp_best_lambda,
                     pred.at = 365 * 2,
                     funlabel = "2-Year Overall Survival Probability")

plot(nom)
}

