% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hddnom-package.R
\docType{package}
\name{hddnom-package}
\alias{hddnom-package}
\title{Nomograms for High-Dimensional Cox Models}
\description{
Nomograms for High-Dimensional Cox Models
}
\details{
The vignette can be opened with \code{vignette('hddnom')}.

\tabular{ll}{ Package: \tab hddnom\cr Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Miaozhu Li <\email{miaozhu.li@duke.edu}>
        Nan Xiao <\email{nanx@uchicago.edu}>
}

