% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{hdnom.calibrate.internal.true}
\alias{hdnom.calibrate.internal.true}
\title{Compute Kaplan-Meier estimated survival probabilities for calibration}
\usage{
hdnom.calibrate.internal.true(pred_prob, grp, time, event, pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier estimated survival probabilities for calibration
}
\keyword{internal}

