% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-2-calibrate-external.R
\name{hdnom.external.calibrate.internal.true}
\alias{hdnom.external.calibrate.internal.true}
\title{Compute Kaplan-Meier Estimated Survival Probabilities for External Calibration}
\usage{
hdnom.external.calibrate.internal.true(pred_prob, grp, time_new, event_new,
  pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier Estimated Survival Probabilities for External Calibration
}
\keyword{internal}
