% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnomo-package.R
\docType{package}
\name{hdnomo-package}
\alias{hdnomo-package}
\title{Nomograms for High-Dimensional Penalized Cox Models}
\description{
Nomograms for High-Dimensional Penalized Cox Models
}
\details{
The vignette can be opened with \code{vignette('hdnomo')}.

\tabular{ll}{ Package: \tab hdnomo\cr Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Miaozhu Li <\email{miaozhu.li@duke.edu}>
        Nan Xiao <\email{nanx@uchicago.edu}>
}

