% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom-datalist.R
\docType{data}
\name{mirna}
\alias{mirna}
\title{TCGA Glioblastoma Multiforme (GBM) mircoRNA expression data}
\usage{
data(mirna)
}
\description{
Processed TCGA Glioblastoma Multiforme (GBM) mircoRNA expression data,
with survival time and status, from Wang et, al. 2014.
}
\details{
The dataset is a list with three components:
\code{x}, \code{time}, and \code{event}.
}
\examples{
data(mirna)
x = mirna$x
time = mirna$time
event = mirna$event
dim(x)
}
\references{
Wang, B. (2014). Similarity network fusion for aggregating data types
on a genomic scale. \emph{Nature methods}, 11(3), 333--337.
}

