% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_survial_curves.R
\name{ncvreg.basesurv}
\alias{ncvreg.basesurv}
\title{Breslow Baseline Hazard Estimator for ncvreg Objects}
\usage{
ncvreg.basesurv(time, event, lp, times.eval = NULL, centered = FALSE)
}
\arguments{
\item{time}{Survival time}

\item{event}{Status indicator}

\item{lp}{Linear predictors}

\item{times.eval}{Survival time to evaluate}

\item{centered}{Should we center the survival curve?
See \code{\link[survival]{basehaz}} for details.}
}
\value{
list containing cumulative base hazard
}
\description{
Derived from \code{peperr:::basesurv} and \code{gbm::basehaz.gbm}.
}
\examples{
NULL
}
