% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{ncvreg.calibrate.internal.pred}
\alias{ncvreg.calibrate.internal.pred}
\title{Compute ncvreg predicted survival probabilities for calibration}
\usage{
ncvreg.calibrate.internal.pred(x_tr, x_te, y_tr, y_te, model.type, alpha,
  lambda, gamma, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute ncvreg predicted survival probabilities for calibration
}
\keyword{internal}

