% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_external.R
\name{ncvreg.external.calibrate.internal.pred}
\alias{ncvreg.external.calibrate.internal.pred}
\title{Compute ncvreg Predicted Survival Probabilities for External Calibration}
\usage{
ncvreg.external.calibrate.internal.pred(object, x_tr, x_te, y_tr, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute ncvreg Predicted Survival Probabilities for External Calibration
}
\keyword{internal}
