% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.nomogram.R
\name{ncvreg.survcurve}
\alias{ncvreg.survcurve}
\title{Survival curve prediction for ncvreg objects}
\usage{
ncvreg.survcurve(object, time, event, x, survtime, lambda)
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\keyword{internal}

