% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.models-mcp.R
\name{ncvreg.tune.gamma}
\alias{ncvreg.tune.gamma}
\title{Automatic MCP/SCAD gamma tuning function by k-fold cross-validation}
\usage{
ncvreg.tune.gamma(..., gammas, seed, parallel)
}
\value{
best model object and best gamma
}
\description{
Automatic MCP/SCAD gamma tuning function by k-fold cross-validation
}
\keyword{internal}

