% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_validate.R
\name{ncvreg.validate.internal}
\alias{ncvreg.validate.internal}
\title{Compute Validation Measures for ncvreg Model Objects}
\usage{
ncvreg.validate.internal(x_tr, x_te, y_tr, y_te, model.type, gamma, alpha,
  lambda, tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute Validation Measures for ncvreg Model Objects
}
\keyword{internal}
