% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_utils.R
\name{ncvreg_calibrate_surv_prob_pred}
\alias{ncvreg_calibrate_surv_prob_pred}
\title{Compute ncvreg predicted survival probabilities for calibration}
\usage{
ncvreg_calibrate_surv_prob_pred(x_tr, x_te, y_tr, model.type, alpha,
  lambda, gamma, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute ncvreg predicted survival probabilities for calibration
}
\keyword{internal}
