% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_model_tune.R
\name{ncvreg_tune_gamma}
\alias{ncvreg_tune_gamma}
\title{Automatic MCP/SCAD gamma tuning function by k-fold cross-validation}
\usage{
ncvreg_tune_gamma(..., gammas, eps, max.iter, seed, parallel)
}
\value{
best model object and best gamma
}
\description{
Automatic MCP/SCAD gamma tuning function by k-fold cross-validation
}
\keyword{internal}
