% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_model_tune.R
\name{ncvreg_tune_gamma_alpha}
\alias{ncvreg_tune_gamma_alpha}
\title{Automatic Mnet/Snet gamma and alpha tuning function by k-fold cross-validation}
\usage{
ncvreg_tune_gamma_alpha(..., gammas, alphas, eps, max.iter, seed, parallel)
}
\value{
best model object, best gamma, and best alpha
}
\description{
Automatic Mnet/Snet gamma and alpha tuning function by k-fold cross-validation
}
\keyword{internal}
