% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_validate_utils.R
\name{ncvreg_validate_tauc}
\alias{ncvreg_validate_tauc}
\title{Compute validation measures for ncvreg model objects}
\usage{
ncvreg_validate_tauc(x_tr, x_te, y_tr, y_te, model.type, gamma, alpha,
  lambda, tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute validation measures for ncvreg model objects
}
\keyword{internal}
