% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11-hdnom-color-palette.R
\name{palette.jco}
\alias{palette.jco}
\title{Color Palette for Journal of Clinical Oncology (JCO)}
\usage{
palette.jco()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in \emph{Journal of Clinical Oncology}.
}
\keyword{internal}

