% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_aaas}
\alias{palette_aaas}
\title{Color Palette for AAAS Journals}
\usage{
palette_aaas()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in AAAS journals such as \emph{Science}.
}
\keyword{internal}
