% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_lancet}
\alias{palette_lancet}
\title{Color Palette for Lancet Journals}
\usage{
palette_lancet()
}
\description{
A 9-color discrete color palette based on the colors
used in figures in \emph{Lancet Oncology}.
}
\keyword{internal}
