% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_npg}
\alias{palette_npg}
\title{Color Palette for NPG Journals}
\usage{
palette_npg()
}
\description{
A 10-color discrete color palette based on the colors
used in figures in NPG journals such as \emph{Nature Reviews Cancer}.
}
\keyword{internal}
