% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{penalized.calibrate.internal.pred}
\alias{penalized.calibrate.internal.pred}
\title{Compute "penalized" predicted survival probabilities for calibration}
\usage{
penalized.calibrate.internal.pred(x_tr, x_te, y_tr, y_te, lambda, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute "penalized" predicted survival probabilities for calibration
}
\keyword{internal}

