% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_external.R
\name{penalized.external.calibrate.internal.pred}
\alias{penalized.external.calibrate.internal.pred}
\title{Compute "penalized" Predicted Survival Probabilities for External Calibration}
\usage{
penalized.external.calibrate.internal.pred(object, x_tr, x_te, y_tr,
  pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute "penalized" Predicted Survival Probabilities for External Calibration
}
\keyword{internal}
