% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.hdnom.external.validate.R
\name{penalized.external.validate.internal}
\alias{penalized.external.validate.internal}
\title{Compute External Validation Measures for "penalized" Model Objects}
\usage{
penalized.external.validate.internal(x_tr, x_te, y_tr, y_te, lambda, tauc.type,
  tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute External Validation Measures for "penalized" Model Objects
}
\keyword{internal}

