% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/penalized.nomogram.R
\name{penalized.nomogram}
\alias{penalized.nomogram}
\title{Nomograms for Cox models fitted with penalized}
\usage{
penalized.nomogram()
}
\description{
Nomograms for Cox models fitted with penalized
}
\examples{
NULL
}

