% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.nomogram.R
\name{penalized.survcurve}
\alias{penalized.survcurve}
\title{Survival curve prediction for "penalized" objects}
\usage{
penalized.survcurve(object, time, event, x, survtime)
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\keyword{internal}

