% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-2-tune.R
\name{penalized.tune.lambda}
\alias{penalized.tune.lambda}
\title{Automatic lambda tuning function for fused lasso by k-fold cross-validation}
\usage{
penalized.tune.lambda(..., lambda1, lambda2, seed, trace, parallel)
}
\value{
best model object, best lambda1, and best lambda2
}
\description{
Automatic lambda tuning function for fused lasso by k-fold cross-validation
}
\keyword{internal}
