% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_5_calibrate_external_utils.R
\name{penalized_calibrate_external_surv_prob_pred}
\alias{penalized_calibrate_external_surv_prob_pred}
\title{Compute penfit predicted survival probabilities for external calibration}
\usage{
penalized_calibrate_external_surv_prob_pred(object, x_tr, x_te, y_tr,
  pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute penfit predicted survival probabilities for external calibration
}
\keyword{internal}
