% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_calibrate_utils.R
\name{penalized_calibrate_surv_prob_pred}
\alias{penalized_calibrate_surv_prob_pred}
\title{Compute penfit predicted survival probabilities for calibration}
\usage{
penalized_calibrate_surv_prob_pred(x_tr, x_te, y_tr, lambda1, lambda2, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute penfit predicted survival probabilities for calibration
}
\keyword{internal}
