% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_2_model_tune.R
\name{penalized_tune_lambda}
\alias{penalized_tune_lambda}
\title{Automatic lambda tuning function for fused lasso by k-fold cross-validation}
\usage{
penalized_tune_lambda(..., lambda1, lambda2, seed, trace, parallel)
}
\value{
best model object, best lambda1, and best lambda2
}
\description{
Automatic lambda tuning function for fused lasso by k-fold cross-validation
}
\keyword{internal}
