% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_validate_utils.R
\name{penalized_validate_tauc}
\alias{penalized_validate_tauc}
\title{Compute validation measures for penfit model objects}
\usage{
penalized_validate_tauc(x_tr, x_te, y_tr, y_te, lambda1, lambda2,
  tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute validation measures for penfit model objects
}
\keyword{internal}
