% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{plot.glmnet.calibrate}
\alias{plot.glmnet.calibrate}
\title{Plot calibration curves}
\usage{
\method{plot}{glmnet.calibrate}(x, ...)
}
\arguments{
\item{x}{a \code{"glmnet.calibrate"} object.}

\item{...}{other parameters for \code{plot}.}
}
\description{
Plot calibration curves
}
\examples{
NULL
}

