% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{plot.glmnet.nomogram}
\alias{plot.glmnet.nomogram}
\title{Plot nomogram objects generated with glmnet.nomogram}
\usage{
\method{plot}{glmnet.nomogram}(x, ...)
}
\arguments{
\item{x}{a \code{"glmnet.nomogram"} object.}

\item{...}{other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Plot nomogram objects generated with glmnet.nomogram
}
\examples{
NULL
}

