% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_6_calibrate_external_method.R
\name{plot.hdnom.calibrate.external}
\alias{plot.hdnom.calibrate.external}
\title{Plot external calibration results}
\usage{
\method{plot}{hdnom.calibrate.external}(x, xlim = c(0, 1), ylim = c(0,
  1), col.pal = c("JCO", "Lancet", "NPG", "AAAS"), ...)
}
\arguments{
\item{x}{An object returned by \code{\link{calibrate_external}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{...}{Other parameters for \code{plot}.}
}
\description{
Plot external calibration results
}
\examples{
NULL
}
