% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.hdnom.compare.validate.R
\name{plot.hdnom.compare.validate}
\alias{plot.hdnom.compare.validate}
\title{Plot Model Comparison by Validation Results}
\usage{
\method{plot}{hdnom.compare.validate}(x, interval = FALSE, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.compare.validate}}.}

\item{interval}{Show maximum, minimum, 0.25, and 0.75 quantiles of
time-dependent AUC as ribbons? Default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot Model Comparison by Validation Results
}
\examples{
NULL
}

