% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_nomogram_method.R
\name{plot.hdnom.nomogram}
\alias{plot.hdnom.nomogram}
\title{Plot nomogram objects}
\usage{
\method{plot}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{as_nomogram}}.}

\item{...}{Other parameters.}
}
\description{
Plot nomogram objects
}
\examples{
NULL
}
