% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_6_validate_external_method.R
\name{plot.hdnom.validate.external}
\alias{plot.hdnom.validate.external}
\title{Plot time-dependent discrimination curves for external validation}
\usage{
\method{plot}{hdnom.validate.external}(x, col.pal = c("JCO", "Lancet", "NPG", "AAAS"), ylim = NULL, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{validate_external}}.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{ylim}{Range of y coordinates. For example, \code{c(0.5, 1)}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot time-dependent discrimination curves for external validation
}
\examples{
NULL
}
