% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{print.glmnet.nomogram}
\alias{print.glmnet.nomogram}
\title{Print nomograms objects generated with glmnet.nomogram}
\usage{
\method{print}{glmnet.nomogram}(x, ...)
}
\arguments{
\item{x}{a \code{"glmnet.nomogram"} object.}

\item{...}{other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Print nomograms objects generated with glmnet.nomogram
}
\examples{
library("glmnet")
library("survival")
library("rms")

# Load imputed SMART data
data(smart)
x = as.matrix(smart[, -c(1, 2)])
time = smart$TEVENT
event = smart$EVENT
x.df = as.data.frame(x)
dd = datadist(x.df)
options(datadist = "dd")

# Fit penalized Cox model (lasso penalty) with glmnet
set.seed(1010)
cvfit = cv.glmnet(x, Surv(time, event), family = "cox", nfolds = 10)
fit = glmnet(x, Surv(time, event), family = "cox")

# Generate glmnet.nomogram objects and draw nomogram
nom = glmnet.nomogram(fit, x, time, event, x.df,
                      s = cvfit$lambda.1se, pred.at = 365 * 2,
                      funlabel = "2-Year Overall Survival Probability")

print(nom)
}

