% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_6_calibrate_external_method.R
\name{print.hdnom.calibrate.external}
\alias{print.hdnom.calibrate.external}
\title{Print external calibration results}
\usage{
\method{print}{hdnom.calibrate.external}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{calibrate_external}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print external calibration results
}
\examples{
NULL
}
