% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{print.hdnom.calibrate}
\alias{print.hdnom.calibrate}
\title{Print calibration result generated by hdnom.calibrate}
\usage{
\method{print}{hdnom.calibrate}(x, ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.calibrate}}.}

\item{...}{other parameters (not used).}
}
\description{
Print calibration result generated by hdnom.calibrate
}
\examples{
NULL
}

