% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_4_compare_by_calibrate_method.R
\name{print.hdnom.compare.calibrate}
\alias{print.hdnom.compare.calibrate}
\title{Print model comparison by calibration results}
\usage{
\method{print}{hdnom.compare.calibrate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{compare_by_calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print model comparison by calibration results
}
\examples{
NULL
}
