% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.hdnom.compare.validate.R
\name{print.hdnom.compare.validate}
\alias{print.hdnom.compare.validate}
\title{Print Model Comparison by Validation Results}
\usage{
\method{print}{hdnom.compare.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.compare.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print Model Comparison by Validation Results
}
\examples{
NULL
}

