% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.nomogram.R
\name{print.hdnom.nomogram}
\alias{print.hdnom.nomogram}
\title{Print nomograms objects generated by hdnom.nomogram}
\usage{
\method{print}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.nomogram}}.}

\item{...}{Other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Print nomograms objects generated by hdnom.nomogram
}
\examples{
NULL
}

