% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_6_validate_external_method.R
\name{print.hdnom.validate.external}
\alias{print.hdnom.validate.external}
\title{Print external validation results}
\usage{
\method{print}{hdnom.validate.external}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{validate_external}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print external validation results
}
\examples{
NULL
}
