% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_3_validate_method.R
\name{print.hdnom.validate}
\alias{print.hdnom.validate}
\title{Print validation results}
\usage{
\method{print}{hdnom.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print validation results
}
\examples{
NULL
}
