% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.predict.R
\name{selected.info}
\alias{selected.info}
\title{Extract Information of Selected Variables from High-Dimensional Cox Models}
\usage{
selected.info(object, x)
}
\arguments{
\item{object}{Model object fitted by \code{hdcox.*()} functions.}

\item{x}{Data matrix used to fit the model.}
}
\description{
Extract the names and type of selected variables from established
high-dimensional Cox models.
}
\examples{
NULL
}

