% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hddnom-datalist.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulation data derived from rms}
\usage{
data(simdata)
}
\description{
Simulation data derived from rms
https://github.com/harrelfe/rms/blob/master/demo/all.R
Detailed Example 1, 2, 3
}
\examples{
data(simdata)
str(simdata)
}

