% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{summary.glmnet.calibrate}
\alias{summary.glmnet.calibrate}
\title{Summary calibration result generated with glmnet.calibrate}
\usage{
\method{summary}{glmnet.calibrate}(object, ...)
}
\arguments{
\item{object}{a \code{"glmnet.calibrate"} object.}

\item{...}{other parameters (not used).}
}
\description{
Summary calibration result generated with glmnet.calibrate
}
\examples{
NULL
}

