% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.validate.R
\name{summary.glmnet.validate}
\alias{summary.glmnet.validate}
\title{Summary validation result generated with glmnet.validate}
\usage{
\method{summary}{glmnet.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{a \code{"glmnet.validate"} object.}

\item{silent}{Print summary table header or not.
Default is \code{FALSE}.}

\item{...}{other parameters (not used).}
}
\description{
Summary validation result generated with glmnet.validate
}
\examples{
NULL
}

