% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_6_calibrate_external_method.R
\name{summary.hdnom.calibrate.external}
\alias{summary.hdnom.calibrate.external}
\title{Summary of external calibration results}
\usage{
\method{summary}{hdnom.calibrate.external}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{calibrate_external}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of external calibration results
}
\examples{
NULL
}
