% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_3_calibrate_method.R
\name{summary.hdnom.calibrate}
\alias{summary.hdnom.calibrate}
\title{Summary of calibration results}
\usage{
\method{summary}{hdnom.calibrate}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of calibration results
}
\examples{
NULL
}
