% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_4_compare_by_calibrate_method.R
\name{summary.hdnom.compare.calibrate}
\alias{summary.hdnom.compare.calibrate}
\title{Summary of model comparison by calibration results}
\usage{
\method{summary}{hdnom.compare.calibrate}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{compare_by_calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of model comparison by calibration results
}
\examples{
NULL
}
