% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_compare_by_validate_method.R
\name{summary.hdnom.compare.validate}
\alias{summary.hdnom.compare.validate}
\title{Summary of model comparison by validation results}
\usage{
\method{summary}{hdnom.compare.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object \code{\link{compare_by_validate}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of model comparison by validation results
}
\examples{
NULL
}
