% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-hdnom-external-calibrate.R
\name{summary.hdnom.external.calibrate}
\alias{summary.hdnom.external.calibrate}
\title{Summary of External Calibration Results}
\usage{
\method{summary}{hdnom.external.calibrate}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{hdnom.external.calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of External Calibration Results
}
\examples{
NULL
}
