% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5.hdnom.external.validate.R
\name{summary.hdnom.external.validate}
\alias{summary.hdnom.external.validate}
\title{Summary of External Validation Results}
\usage{
\method{summary}{hdnom.external.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object \code{\link{hdnom.external.validate}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of External Validation Results
}
\examples{
NULL
}

