% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_6_validate_external_method.R
\name{summary.hdnom.validate.external}
\alias{summary.hdnom.validate.external}
\title{Summary of external validation results}
\usage{
\method{summary}{hdnom.validate.external}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{validate_external}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of external validation results
}
\examples{
NULL
}
