% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{summary.hdnom.validate}
\alias{summary.hdnom.validate}
\title{Summary validation result generated by hdnom.validate}
\usage{
\method{summary}{hdnom.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object \code{\link{hdnom.validate}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary validation result generated by hdnom.validate
}
\examples{
NULL
}

