% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{survcurve.glmnet}
\alias{survcurve.glmnet}
\title{Survival curve prediction for glmnet objects}
\usage{
survcurve.glmnet(object, time, event, x, survtime, lambda)
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\keyword{internal}

