% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hdnom.R
\name{theme_hdnom}
\alias{theme_hdnom}
\title{Plot theme (ggplot2) for hdnom}
\usage{
theme_hdnom(base_size = 14)
}
\arguments{
\item{base_size}{base font size}
}
\description{
Plot theme (ggplot2) for hdnom
}
