% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.models.R
\name{tuneAlpha}
\alias{tuneAlpha}
\title{Automatic alpha tuning function by k-fold cross-validation}
\usage{
tuneAlpha(..., alphas, seed)
}
\value{
best model object and best alpha
}
\description{
Automatic alpha tuning function by k-fold cross-validation
}
\keyword{internal}

