% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{glmnet.basesurv}
\alias{glmnet.basesurv}
\title{Breslow Baseline Hazard Estimator for glmnet Objects}
\usage{
glmnet.basesurv(time, event, lp, times.eval = NULL, centered = FALSE)
}
\arguments{
\item{time}{Survival time}

\item{event}{Status indicator}

\item{lp}{Linear predictors}

\item{times.eval}{Survival time to evaluate}

\item{centered}{Should we center the survival curve?
See [survival::basehaz()] for details.}
}
\value{
list containing cumulative base hazard
}
\description{
Derived from \code{peperr:::basesurv} and \code{gbm::basehaz.gbm}.
}
\examples{
NULL
}
