% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{glmnet.calibrate.internal.pred}
\alias{glmnet.calibrate.internal.pred}
\title{Compute glmnet Predicted Survival Probabilities for Calibration}
\usage{
glmnet.calibrate.internal.pred(x_tr, x_te, y_tr, y_te, alpha, lambda,
  pen.factor, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute glmnet Predicted Survival Probabilities for Calibration
}
\keyword{internal}

