% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{glmnet.calibrate}
\alias{glmnet.calibrate}
\title{Calibrate Cox models fitted with glmnet}
\usage{
glmnet.calibrate()
}
\description{
Calibrate Cox models fitted with glmnet
}
\examples{
set.seed(1003)
}

