% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{glmnet.survcurve}
\alias{glmnet.survcurve}
\title{Survival Curve Prediction for glmnet Objects}
\usage{
glmnet.survcurve(object, time, event, x, survtime)
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\keyword{internal}
