% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-hdnom-models.R
\name{glmnet.tune.alpha}
\alias{glmnet.tune.alpha}
\title{Automatic alpha tuning function by k-fold cross-validation}
\usage{
glmnet.tune.alpha(..., alphas, seed, parallel)
}
\value{
best model object and best alpha
}
\description{
Automatic alpha tuning function by k-fold cross-validation
}
\keyword{internal}

