% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{glmnet.validate.internal}
\alias{glmnet.validate.internal}
\title{Compute Validation Measures for glmnet Objects}
\usage{
glmnet.validate.internal(x_tr, x_te, y_tr, y_te, alpha, lambda, pen.factor,
  tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute Validation Measures for glmnet Objects
}
\keyword{internal}

