% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.models-scad.R
\name{hdcox.snet}
\alias{hdcox.snet}
\title{Snet model selection for high-dimensional Cox models}
\usage{
hdcox.snet(x, y, nfolds = 5L, gammas = c(2.01, 2.3, 3.7, 200),
  alphas = seq(0.05, 0.95, 0.05), seed = 1001, trace = FALSE,
  parallel = FALSE)
}
\arguments{
\item{x}{Data matrix.}

\item{y}{Response matrix made by \code{\link[survival]{Surv}}.}

\item{nfolds}{Fold numbers of cross-validation.}

\item{gammas}{Gammas to tune in \code{\link[ncvreg]{cv.ncvsurv}}.}

\item{alphas}{Alphas to tune in \code{\link[ncvreg]{cv.ncvsurv}}.}

\item{seed}{A random seed for cross-validation fold division.}

\item{trace}{Output the cross-validation parameter tuning
progress or not. Default is \code{FALSE}.}
}
\description{
Automatic Snet model selection for high-dimensional
Cox models, evaluated by penalized partial-likelihood.
}
\examples{
library("ncvreg")
library("survival")
library("rms")

# Load imputed SMART data
data("smart")
x = as.matrix(smart[, -c(1, 2)])
time = smart$TEVENT
event = smart$EVENT
y = Surv(time, event)

# Fit Cox model by Snet penalization
snetfit = hdcox.snet(x, y, nfolds = 3, gammas = c(2.01, 3.7),
                     alphas = c(0.3, 0.8), trace = TRUE)

# Prepare data for hdnom.nomogram
x.df = as.data.frame(x)
dd = datadist(x.df)
options(datadist = "dd")

# Generate hdnom.nomogram objects and plot nomogram
nom = hdnom.nomogram(snetfit$snet_model, x, time, event, x.df,
                     gamma = snetfit$snet_best_gamma,
                     alpha = snetfit$snet_best_alpha,
                     lambda = snetfit$snet_best_lambda,
                     pred.at = 365 * 2,
                     funlabel = "2-Year Overall Survival Probability")

plot(nom)
}

