% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom-package.R
\docType{package}
\name{hdnom-package}
\alias{hdnom-package}
\title{Benchmarking and Visualization Toolkit for Penalized Cox Models}
\description{
Benchmarking and Visualization Toolkit for Penalized Cox Models
}
\details{
Browse vignettes with \code{browseVignettes(package = "hdnom")}.

\tabular{ll}{ Package: \tab hdnom\cr
Type: \tab Package\cr
License: \tab GPL-3\cr }
}
\author{
Nan Xiao <\email{me@nanx.me}>
        Qingsong Xu <\email{qsxu@csu.edu.cn}>
        Miaozhu Li <\email{miaozhu.li@duke.edu}>
}
