% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom-package.R
\docType{package}
\name{hdnom-package}
\alias{hdnom-package}
\title{Nomograms for High-Dimensional Data with Penalized Cox Models}
\description{
Nomograms for High-Dimensional Data with Penalized Cox Models
}
\details{
The vignette can be opened with \code{vignette('hdnom')}.

\tabular{ll}{ Package: \tab hdnom\cr
Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Miaozhu Li <\email{miaozhu.li@duke.edu}>
        Nan Xiao <\email{me@nanx.me}>
}

