% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom-package.R
\docType{package}
\name{hdnom-package}
\alias{hdnom-package}
\title{Benchmarking and Visualization Toolkit for Penalized Cox Models}
\description{
Browse vignettes with \code{browseVignettes(package = "hdnom")}.
}
\details{
\tabular{ll}{
Package: \tab hdnom\cr
Type: \tab Package\cr
License: \tab GPL-3}
}
