% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1-calibrate.R
\name{hdnom.calibrate.internal.true}
\alias{hdnom.calibrate.internal.true}
\title{Compute Kaplan-Meier Estimated Survival Probabilities for Calibration}
\usage{
hdnom.calibrate.internal.true(pred_prob, grp, time, event, pred.at, ngroup)
}
\value{
list
}
\description{
Compute Kaplan-Meier Estimated Survival Probabilities for Calibration
}
\keyword{internal}
