% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.hdnom.kmplot.R
\name{kmplot}
\alias{kmplot}
\title{Kaplan-Meier Plot with Number at Risk Table}
\usage{
kmplot(fit, group.name = NULL, time.at = NULL, surv.df = NULL)
}
\arguments{
\item{fit}{a \code{\link[survival]{survfit}} object}

\item{group.name}{Group labels. Default is Group 1, Group 2, ... Group k.}

\item{time.at}{Time points to evaluate the number at risk.}

\item{surv.df}{Data frame containing survival time, event and risk group
for log-rank test.}
}
\description{
Kaplan-Meier Plot with Number at Risk Table
}
\keyword{internal}

