% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{ncvreg.basesurv}
\alias{ncvreg.basesurv}
\title{Breslow Baseline Hazard Estimator for ncvreg Objects}
\usage{
ncvreg.basesurv(time, event, lp, times.eval = NULL, centered = FALSE)
}
\value{
list containing cumulative base hazard
}
\description{
Derived from \code{peperr:::basesurv} and \code{gbm::basehaz.gbm}.
}
\keyword{internal}
