% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-hdnom-calibrate.R
\name{ncvreg.calibrate.internal.pred}
\alias{ncvreg.calibrate.internal.pred}
\title{Compute ncvreg Predicted Survival Probabilities for Calibration}
\usage{
ncvreg.calibrate.internal.pred(x_tr, x_te, y_tr, model.type, alpha, lambda,
  gamma, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute ncvreg Predicted Survival Probabilities for Calibration
}
\keyword{internal}

