% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.hdnom.external.validate.R
\name{ncvreg.external.validate.internal}
\alias{ncvreg.external.validate.internal}
\title{Compute External Validation Measures for ncvreg Model Objects}
\usage{
ncvreg.external.validate.internal(x_tr, x_te, y_tr, y_te, model.type, gamma,
  alpha, lambda, tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute External Validation Measures for ncvreg Model Objects
}
\keyword{internal}

