% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-3-survial-curves.R
\name{ncvreg.survcurve}
\alias{ncvreg.survcurve}
\title{Survival Curve Prediction for ncvreg Objects}
\usage{
ncvreg.survcurve(object, time, event, x, survtime)
}
\arguments{
\item{object}{\code{ncvreg} model object}

\item{time}{Survival time}

\item{event}{Status indicator}

\item{x}{Predictor matrix}

\item{survtime}{Survival time to evaluate}
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\examples{
NULL
}
