% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{ncvreg.survcurve}
\alias{ncvreg.survcurve}
\title{Survival Curve Prediction for ncvreg Objects}
\usage{
ncvreg.survcurve(object, time, event, x, survtime)
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\keyword{internal}

