% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-1-models.R
\name{ncvreg.tune.gamma.alpha}
\alias{ncvreg.tune.gamma.alpha}
\title{Automatic Mnet/Snet gamma and alpha tuning function by k-fold cross-validation}
\usage{
ncvreg.tune.gamma.alpha(..., gammas, alphas, eps, max.iter, seed, parallel)
}
\value{
best model object, best gamma, and best alpha
}
\description{
Automatic Mnet/Snet gamma and alpha tuning function by k-fold cross-validation
}
\keyword{internal}
