% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{ncvreg.validate.internal}
\alias{ncvreg.validate.internal}
\title{Compute validation measures for ncvreg model objects}
\usage{
ncvreg.validate.internal(x_tr, x_te, y_tr, y_te, model.type, gamma, alpha,
  lambda, tauc.type, tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute validation measures for ncvreg model objects
}
\keyword{internal}

