% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-hdnom-calibrate.R
\name{penalized.calibrate.internal.pred}
\alias{penalized.calibrate.internal.pred}
\title{Compute "penalized" Predicted Survival Probabilities for Calibration}
\usage{
penalized.calibrate.internal.pred(x_tr, x_te, y_tr, lambda1, lambda2, pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute "penalized" Predicted Survival Probabilities for Calibration
}
\keyword{internal}

