% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.hdnom.external.calibrate.R
\name{penalized.external.calibrate.internal.pred}
\alias{penalized.external.calibrate.internal.pred}
\title{Compute "penalized" Predicted Survival Probabilities for External Calibration}
\usage{
penalized.external.calibrate.internal.pred(x_tr, x_te, y_tr, y_te, lambda,
  pred.at)
}
\value{
list containing predicted survival probability
}
\description{
Compute "penalized" Predicted Survival Probabilities for External Calibration
}
\keyword{internal}

