% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{penalized.survcurve}
\alias{penalized.survcurve}
\title{Survival Curve Prediction for "penalized" Objects}
\usage{
penalized.survcurve(object, time, event, x, survtime)
}
\arguments{
\item{object}{\code{penalized} model object}

\item{time}{Survival time}

\item{event}{Status indicator}

\item{x}{Predictor matrix}

\item{survtime}{Survival time to evaluate}
}
\value{
list containing predicted survival probabilities and
linear predictors for all samples
}
\description{
Derived from c060::predictProb.coxnet
}
\examples{
NULL
}
