% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-hdnom-validate.R
\name{penalized.validate.internal}
\alias{penalized.validate.internal}
\title{Compute Validation Measures for "penalized" Model Objects}
\usage{
penalized.validate.internal(x_tr, x_te, y_tr, y_te, lambda1, lambda2, tauc.type,
  tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute Validation Measures for "penalized" Model Objects
}
\keyword{internal}
