% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{penalized.validate.internal}
\alias{penalized.validate.internal}
\title{Compute validation measures for "penalized" model objects}
\usage{
penalized.validate.internal(x_tr, x_te, y_tr, y_te, lambda, tauc.type,
  tauc.time)
}
\value{
time-dependent AUC (tAUC) value
}
\description{
Compute validation measures for "penalized" model objects
}
\keyword{internal}

