% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.calibrate.R
\name{plot.glmnet.calibrate}
\alias{plot.glmnet.calibrate}
\title{Plot calibration result objects generated with glmnet.calibrate}
\usage{
\method{plot}{glmnet.calibrate}(object)
}
\arguments{
\item{object}{\code{"glmnet.calibrate"} object.}
}
\description{
Plot calibration result objects generated with glmnet.calibrate
}
\examples{
NULL
}

