% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{plot.glmnet.nomogram}
\alias{plot.glmnet.nomogram}
\title{Plot Nomograms Objects Generated with glmnet.nomogram}
\usage{
\method{plot}{glmnet.nomogram}(object)
}
\arguments{
\item{object}{\code{"glmnet.nomogram"} object.}
}
\description{
Plot Nomograms Objects Generated with glmnet.nomogram
}
\examples{
library("glmnet")
library("survival")
library("rms")

# Load imputed SMART data
data(smart)
x = as.matrix(smart[, -c(1, 2)])
time = smart$TEVENT
event = smart$EVENT
x.df = as.data.frame(x)
dd = datadist(x.df)
options(datadist = "dd")

# Fit penalized Cox models with glmnet
set.seed(1010)
cvfit = cv.glmnet(x, Surv(time, event), family = "cox", nfolds = 10)
fit = glmnet(x, Surv(time, event), family = "cox")

# Generate glmnet.nomogram objects and draw nomogram
nom = glmnet.nomogram(fit, x, time, event, x.df,
                      s = cvfit$lambda.1se, pred.at = 365 * 2,
                      funlabel = "2-Year Overall Survival Probability")

plot(nom)
}

