% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.validate.R
\name{plot.glmnet.validate}
\alias{plot.glmnet.validate}
\title{Plot validation result generated with glmnet.validate}
\usage{
\method{plot}{glmnet.validate}(x, ...)
}
\arguments{
\item{x}{a \code{"glmnet.validate"} object.}

\item{...}{other parameters for \code{plot}.}
}
\description{
Plot validation result generated with glmnet.validate
}
\examples{
NULL
}

