% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.hdnom.calibrate.R
\name{plot.hdnom.calibrate}
\alias{plot.hdnom.calibrate}
\title{Plot Calibration Results}
\usage{
\method{plot}{hdnom.calibrate}(x, xlim = c(0, 1), ylim = c(0, 1), ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.calibrate}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{...}{other parameters for \code{plot}.}
}
\description{
Plot Calibration Results
}
\examples{
NULL
}

