% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-1-calibrate.R
\name{plot.hdnom.calibrate}
\alias{plot.hdnom.calibrate}
\title{Plot Calibration Results}
\usage{
\method{plot}{hdnom.calibrate}(x, xlim = c(0, 1), ylim = c(0, 1),
  col.pal = c("JCO", "Lancet", "NPG", "AAAS"), ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.calibrate}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{...}{Other parameters for \code{plot}.}
}
\description{
Plot Calibration Results
}
\examples{
NULL
}
