% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10.hdnom.compare.calibrate.R
\name{plot.hdnom.compare.calibrate}
\alias{plot.hdnom.compare.calibrate}
\title{Plot Model Comparison by Calibration Results}
\usage{
\method{plot}{hdnom.compare.calibrate}(x, xlim = c(0, 1), ylim = c(0, 1),
  ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.compare.calibrate}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{...}{other parameters (not used).}
}
\description{
Plot Model Comparison by Calibration Results
}
\examples{
NULL
}

