% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_4_compare_by_calibrate_method.R
\name{plot.hdnom.compare.calibrate}
\alias{plot.hdnom.compare.calibrate}
\title{Plot model comparison by calibration results}
\usage{
\method{plot}{hdnom.compare.calibrate}(
  x,
  xlim = c(0, 1),
  ylim = c(0, 1),
  col.pal = c("JCO", "Lancet", "NPG", "AAAS"),
  ...
)
}
\arguments{
\item{x}{An object returned by \code{\link{compare_by_calibrate}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot model comparison by calibration results
}
\examples{
NULL
}
