% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_compare_by_validate_method.R
\name{plot.hdnom.compare.validate}
\alias{plot.hdnom.compare.validate}
\title{Plot model comparison by validation results}
\usage{
\method{plot}{hdnom.compare.validate}(
  x,
  interval = FALSE,
  col.pal = c("JCO", "Lancet", "NPG", "AAAS"),
  ylim = NULL,
  ...
)
}
\arguments{
\item{x}{An object returned by \code{\link{compare_by_validate}}.}

\item{interval}{Show maximum, minimum, 0.25, and 0.75 quantiles of
time-dependent AUC as ribbons? Default is \code{FALSE}.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{ylim}{Range of y coordinates. For example, \code{c(0.5, 1)}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot model comparison by validation results
}
\examples{
NULL
}
