% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.hdnom.external.calibrate.R
\name{plot.hdnom.external.calibrate}
\alias{plot.hdnom.external.calibrate}
\title{Plot External Calibration Results}
\usage{
\method{plot}{hdnom.external.calibrate}(x, xlim = c(0, 1), ylim = c(0, 1),
  ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.external.calibrate}}.}

\item{xlim}{x axis limits of the plot.}

\item{ylim}{y axis limits of the plot.}

\item{...}{other parameters for \code{plot}.}
}
\description{
Plot External Calibration Results
}
\examples{
NULL
}

