% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.hdnom.external.validate.R
\name{plot.hdnom.external.validate}
\alias{plot.hdnom.external.validate}
\title{Plot Time-Dependent Discrimination Curves for External Validation}
\usage{
\method{plot}{hdnom.external.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.external.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot Time-Dependent Discrimination Curves for External Validation
}
\examples{
NULL
}

