% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-2-validate-external.R
\name{plot.hdnom.external.validate}
\alias{plot.hdnom.external.validate}
\title{Plot Time-Dependent Discrimination Curves for External Validation}
\usage{
\method{plot}{hdnom.external.validate}(x, col.pal = c("JCO", "Lancet",
  "NPG", "AAAS"), ylim = NULL, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.external.validate}}.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{ylim}{Range of y coordinates. For example, \code{c(0.5, 1)}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot Time-Dependent Discrimination Curves for External Validation
}
\examples{
NULL
}
