% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2.hdnom.nomogram.R
\name{plot.hdnom.nomogram}
\alias{plot.hdnom.nomogram}
\title{Plot Nomogram Objects Generated by hdnom.nomogram}
\usage{
\method{plot}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.nomogram}}.}

\item{...}{Other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Plot Nomogram Objects Generated by hdnom.nomogram
}
\examples{
NULL
}

