% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.nomogram.R
\name{plot.hdnom.nomogram}
\alias{plot.hdnom.nomogram}
\title{Plot nomogram objects generated by hdnom.nomogram}
\usage{
\method{plot}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.nomogram}}.}

\item{...}{Other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Plot nomogram objects generated by hdnom.nomogram
}
\examples{
NULL
}

