% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{plot.hdnom.validate}
\alias{plot.hdnom.validate}
\title{Plot optimism-corrected time-dependent discrimination curves}
\usage{
\method{plot}{hdnom.validate}(x, ylim = c(0.5, 1), xaxt.label = NULL,
  xlab = "Time", ylab = "Time-dependent Area under ROC", ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.validate}}.}

\item{ylim}{The y axis limits of the plot.}

\item{xaxt.label}{The x axis tick label.}

\item{xlab}{Title for the x axis.}

\item{ylab}{Title for the y axis.}

\item{...}{Other parameters for \code{\link{plot}}.}
}
\description{
Plot optimism-corrected time-dependent discrimination curves
}
\examples{
NULL
}

