% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.hdnom.validate.R
\name{plot.hdnom.validate}
\alias{plot.hdnom.validate}
\title{Plot Optimism-Corrected Time-Dependent Discrimination Curves for Validation}
\usage{
\method{plot}{hdnom.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot Optimism-Corrected Time-Dependent Discrimination Curves for Validation
}
\examples{
NULL
}

