% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-hdnom-validate.R
\name{plot.hdnom.validate}
\alias{plot.hdnom.validate}
\title{Plot Optimism-Corrected Time-Dependent Discrimination Curves for Validation}
\usage{
\method{plot}{hdnom.validate}(x, col.pal = c("JCO", "Lancet", "NPG", "AAAS"),
  ylim = NULL, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.validate}}.}

\item{col.pal}{Color palette to use. Possible values are
\code{"JCO"}, \code{"Lancet"}, \code{"NPG"}, and \code{"AAAS"}.
Default is \code{"JCO"}.}

\item{ylim}{Range of y coordinates. For example, \code{c(0.5, 1)}.}

\item{...}{Other parameters (not used).}
}
\description{
Plot Optimism-Corrected Time-Dependent Discrimination Curves for Validation
}
\examples{
NULL
}

