% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glmnet.nomogram.R
\name{print.glmnet.nomogram}
\alias{print.glmnet.nomogram}
\title{Print nomograms objects generated with glmnet.nomogram}
\usage{
\method{print}{glmnet.nomogram}(x, ...)
}
\arguments{
\item{x}{a \code{"glmnet.nomogram"} object.}

\item{...}{other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Print nomograms objects generated with glmnet.nomogram
}
\examples{
NULL
}

