% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_3_calibrate_method.R
\name{print.hdnom.calibrate}
\alias{print.hdnom.calibrate}
\title{Print calibration results}
\usage{
\method{print}{hdnom.calibrate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print calibration results
}
\examples{
NULL
}
