% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-hdnom-calibrate.R
\name{print.hdnom.calibrate}
\alias{print.hdnom.calibrate}
\title{Print Calibration Results}
\usage{
\method{print}{hdnom.calibrate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print Calibration Results
}
\examples{
NULL
}

