% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdnom.compare.calibrate.R
\name{print.hdnom.compare.calibrate}
\alias{print.hdnom.compare.calibrate}
\title{Print Model Comparison by Calibration Results}
\usage{
\method{print}{hdnom.compare.calibrate}(x, ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.compare.calibrate}}.}

\item{...}{other parameters (not used).}
}
\description{
Print Model Comparison by Calibration Results
}
\examples{
NULL
}

