% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_compare_by_validate_method.R
\name{print.hdnom.compare.validate}
\alias{print.hdnom.compare.validate}
\title{Print model comparison by validation results}
\usage{
\method{print}{hdnom.compare.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{compare_by_validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print model comparison by validation results
}
\examples{
NULL
}
