% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9.hdnom.external.calibrate.R
\name{print.hdnom.external.calibrate}
\alias{print.hdnom.external.calibrate}
\title{Print External Calibration Results Generated by hdnom.external.calibrate}
\usage{
\method{print}{hdnom.external.calibrate}(x, ...)
}
\arguments{
\item{x}{an object returned by \code{\link{hdnom.external.calibrate}}.}

\item{...}{other parameters (not used).}
}
\description{
Print External Calibration Results Generated by hdnom.external.calibrate
}
\examples{
NULL
}

