% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6.hdnom.external.calibrate.R
\name{print.hdnom.external.calibrate}
\alias{print.hdnom.external.calibrate}
\title{Print External Calibration Results}
\usage{
\method{print}{hdnom.external.calibrate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.external.calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print External Calibration Results
}
\examples{
NULL
}

