% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_validate_external.R
\name{print.hdnom.external.validate}
\alias{print.hdnom.external.validate}
\title{Print External Validation Results}
\usage{
\method{print}{hdnom.external.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.external.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print External Validation Results
}
\examples{
NULL
}
