% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_4_nomogram_method.R
\name{print.hdnom.nomogram}
\alias{print.hdnom.nomogram}
\title{Print nomograms objects}
\usage{
\method{print}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{as_nomogram}}.}

\item{...}{Other parameters.}
}
\description{
Print nomograms objects
}
\examples{
NULL
}
