% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-hdnom-nomogram.R
\name{print.hdnom.nomogram}
\alias{print.hdnom.nomogram}
\title{Print Nomograms Objects Generated by hdnom.nomogram}
\usage{
\method{print}{hdnom.nomogram}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.nomogram}}.}

\item{...}{Other parameters for \code{\link[rms]{nomogram}}.}
}
\description{
Print Nomograms Objects Generated by hdnom.nomogram
}
\examples{
NULL
}

