% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.validate.R
\name{print.hdnom.validate}
\alias{print.hdnom.validate}
\title{Print validation result generated by hdnom.validate}
\usage{
\method{print}{hdnom.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print validation result generated by hdnom.validate
}
\examples{
NULL
}

