% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_validate.R
\name{print.hdnom.validate}
\alias{print.hdnom.validate}
\title{Print Validation Results}
\usage{
\method{print}{hdnom.validate}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{hdnom.validate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Print Validation Results
}
\examples{
NULL
}
