% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdnom.calibrate.R
\name{summary.hdnom.calibrate}
\alias{summary.hdnom.calibrate}
\title{Summary calibration result generated by hdnom.calibrate}
\usage{
\method{summary}{hdnom.calibrate}(object, ...)
}
\arguments{
\item{object}{an object returned by \code{\link{hdnom.calibrate}}.}

\item{...}{other parameters (not used).}
}
\description{
Summary calibration result generated by hdnom.calibrate
}
\examples{
NULL
}

