% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.hdnom.calibrate.R
\name{summary.hdnom.calibrate}
\alias{summary.hdnom.calibrate}
\title{Summary of Calibration Results Generated by hdnom.calibrate}
\usage{
\method{summary}{hdnom.calibrate}(object, ...)
}
\arguments{
\item{object}{an object returned by \code{\link{hdnom.calibrate}}.}

\item{...}{other parameters (not used).}
}
\description{
Summary of Calibration Results Generated by hdnom.calibrate
}
\examples{
NULL
}

