% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-hdnom-calibrate.R
\name{summary.hdnom.calibrate}
\alias{summary.hdnom.calibrate}
\title{Summary of Calibration Results}
\usage{
\method{summary}{hdnom.calibrate}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{hdnom.calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of Calibration Results
}
\examples{
NULL
}

