% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-2-compare-calibrate.R
\name{summary.hdnom.compare.calibrate}
\alias{summary.hdnom.compare.calibrate}
\title{Summary of Model Comparison by Calibration Results}
\usage{
\method{summary}{hdnom.compare.calibrate}(object, ...)
}
\arguments{
\item{object}{An object returned by \code{\link{hdnom.compare.calibrate}}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of Model Comparison by Calibration Results
}
\examples{
NULL
}
