% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-hdnom-compare-validate.R
\name{summary.hdnom.compare.validate}
\alias{summary.hdnom.compare.validate}
\title{Summary of Model Comparison by Validation Results}
\usage{
\method{summary}{hdnom.compare.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object \code{\link{hdnom.compare.validate}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of Model Comparison by Validation Results
}
\examples{
NULL
}
