% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_3_validate_method.R
\name{summary.hdnom.validate}
\alias{summary.hdnom.validate}
\title{Summary of validation results}
\usage{
\method{summary}{hdnom.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{A \code{\link{validate}} object.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of validation results
}
\examples{
NULL
}
