% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3.hdnom.validate.R
\name{summary.hdnom.validate}
\alias{summary.hdnom.validate}
\title{Summary of Validation Results}
\usage{
\method{summary}{hdnom.validate}(object, silent = FALSE, ...)
}
\arguments{
\item{object}{An object \code{\link{hdnom.validate}}.}

\item{silent}{Print summary table header or not,
default is \code{FALSE}.}

\item{...}{Other parameters (not used).}
}
\description{
Summary of Validation Results
}
\examples{
NULL
}

